fix: changes in ticketsystem

+++ backend/src/main/java/no/hiof/setgroup7/ticketsys/TicketPrice.java
@@ -1,6 +1,6 @@
 package no.hiof.setgroup7.ticketsys;
 
 public interface TicketPrice {
-    double basePrice();
+    int basePrice();
 
 }
